requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
langchain-text-splitters>=0.2.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
ollama>=0.2.0
plotly>=5.20.0
scikit-learn>=1.4.0
numpy>=1.26.0

# Testing
pytest>=9.0.0
pytest-html>=4.2.0
